refactor(SeatUi): Seat to Ui mapper 함수 사용
