+++ app/src/main/java/woowacourse/movie/uimodel/Mapper.kt
@@ -3,8 +3,10 @@ package woowacourse.movie.utils
 import movie.data.MovieDetail
 import movie.data.MovieSchedule
 import movie.screening.ScreeningDate
+import movie.seat.Seat
 import woowacourse.movie.uimodel.MovieDetailUi
 import woowacourse.movie.uimodel.MovieModelUi
+import woowacourse.movie.uimodel.SeatUi
 
 fun MovieModelUi.MovieScheduleUi.toDomain(): MovieSchedule {
     return MovieSchedule(
@@ -17,4 +19,8 @@ fun MovieModelUi.MovieScheduleUi.toDomain(): MovieSchedule {
 
 fun MovieDetailUi.toDomain(): MovieDetail {
     return MovieDetail(count, title, date, time)
+}
+
+fun Seat.toUi(): SeatUi {
+    return SeatUi(row, col)
 }
